+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 * [#80]: Add `disabled` support for elements in the `options` property
 * [#87]: Add `title` support for elements in the `options` property 
 * [#90]: Add `selection` argument to the `onChange` handler to track highlighted values
+* [#104]: Add `className` property to allow specification of a custom class on the root node
 
 ### Bug Fixes
 
 * [#103]: Fix issue where `allowDuplicates` would fail to add extra items when `simpleValue={false}`
+* [#110]: Fix issue where `allowDuplicates` would cause the `available` property to be ignored
 
 ## [v2.0.0](https://github.com/jakezatecky/react-dual-listbox/compare/v1.4.2...v2.0.0) (2019-05-14)